Dashboard for Visualization
The dashboard was created using Brazilian's National Agency for Petroleum (ANP) public information. After a general analysis of its numbers, the price variable was analyzed with the region variable in order to develop a linear model. I started comparing the the data, but the linear model is still to be developed. 
Once it is done, I'll post the lm code as another file.